Moved ApiController out of V3 namespace
